refactor(deno/booking-api): use DSQL postgres.js connector
Replace the raw deno-postgres driver and hand-written IAM token
generator with the official @aws/aurora-dsql-postgresjs-connector.
Rewrite the booking API around postgres.js tagged-template queries
and sql.begin() transactions.

Rename deno/booking-api/deno-postgres/ to deno/booking-api/postgres-js/
to match the other postgres.js sample. Add a CLEANUP_ON_EXIT env
var, URL validation for test-api.sh, and updated README.
