+++ README.md
@@ -61,7 +61,7 @@ The subdirectories contain code examples for connecting and using Aurora DSQL in
 |                          Sample Applications                          |
 |:---------------------------------------------------------------------:|
 | [Amazon Aurora DSQL Agent](sample-amazon-aurora-dsql-agent) — A sample application showing how to build an AI agent that queries Aurora DSQL using natural language, powered by Strands Agents, Bedrock AgentCore Gateway (MCP), and AgentCore Runtime (A2A) |
-| [Booking API (Deno)](deno/booking-api/postgres-js) — A booking/reservation REST API built with Deno.serve() and postgres.js (via the Aurora DSQL connector), demonstrating OCC retry handling, unique-window enforcement, IAM authentication, and serverless-ready architecture |
+| [Booking API (Deno)](deno/booking-api/postgres-js) — A sample REST API for room/resource bookings built with Deno.serve() and postgres.js via the Aurora DSQL connector. Includes OCC retry handling, unique-window enforcement via an async index, and pooled IAM token auth |
 
 Each example includes language and client-specific instructions as well as instructions to invoke example code.
