--- deno/booking-api/deno-postgres/db.ts
@@ -1,89 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-/**
- * Database Connection Module for Aurora DSQL (deno-postgres)
- *
- * Creates per-request connections to an Aurora DSQL cluster using the native
- * `deno-postgres` driver (`jsr:@db/postgres`). Each connection authenticates
- * with a short-lived IAM token generated by the shared token-generator module.
- *
- * Connections use SSL/TLS in verify-full mode via Deno's built-in TLS stack,
- * which automatically uses the system CA certificate store. Aurora DSQL
- * requires SSL — unencrypted connections are rejected.
- *
- * Per-request connections (rather than a pool) are intentional: they make the
- * application serverless-ready, since connection pools don't survive ephemeral
- * instance lifecycles on platforms like Deno Deploy.
- *
- * @module db
- */
-
-import { Client } from "@db/postgres";
-import { generateToken } from "./token-generator.ts";
-
-/**
- * Options for creating a database connection.
- *
- * @property endpoint - The Aurora DSQL cluster hostname
- *   (e.g., "foo0bar1baz2quux3quuux4.dsql.us-east-1.on.aws")
- * @property region - AWS region for the cluster. Passed through to the token
- *   generator, which defaults to `AWS_REGION` env var then "us-east-1".
- * @property user - The PostgreSQL user to connect as (e.g., "admin")
- * @property isAdmin - Whether to generate an admin-level IAM token. Admin
- *   tokens grant full privileges; non-admin tokens are for restricted roles.
- */
-export interface ConnectionOptions {
-  endpoint: string;
-  region?: string;
-  user: string;
-  isAdmin?: boolean;
-}
-
-/**
- * Creates a new connection to an Aurora DSQL cluster.
- *
- * Generates a fresh IAM token, opens a TLS-encrypted connection on port 5432,
- * and returns the connected client. The caller is responsible for closing the
- * connection when done (via `client.end()`).
- *
- * @param options - Connection configuration (endpoint, user, region, admin flag)
- * @returns A connected `Client` instance ready for queries
- * @throws {Error} If token generation or the database connection fails
- *
- * @example
- * ```ts
- * const client = await createConnection({
- *   endpoint: "my-cluster.dsql.us-east-1.on.aws",
- *   user: "admin",
- *   isAdmin: true,
- * });
- * try {
- *   const result = await client.queryObject("SELECT 1 AS ok");
- *   console.log(result.rows);
- * } finally {
- *   await client.end();
- * }
- * ```
- */
-export async function createConnection(
-  options: ConnectionOptions,
-): Promise<Client> {
-  const token = await generateToken({
-    endpoint: options.endpoint,
-    region: options.region,
-    isAdmin: options.isAdmin,
-  });
-
-  const client = new Client({
-    hostname: options.endpoint,
-    port: 5432,
-    user: options.user,
-    password: token,
-    database: "postgres",
-    tls: { enabled: true, enforce: true },
-  });
-
-  await client.connect();
-  return client;
-}
