--- deno/booking-api/deno-postgres/token-generator.ts
@@ -1,79 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-/**
- * IAM Token Generator for Aurora DSQL
- *
- * Generates short-lived IAM authentication tokens for connecting to Aurora DSQL
- * clusters. Aurora DSQL does not use traditional database passwords — instead,
- * it relies on IAM tokens generated by `@aws-sdk/dsql-signer`. Each token is
- * scoped to a specific cluster endpoint and expires after a short period.
- *
- * This module is shared by both booking API driver variants (deno-postgres and
- * node-postgres). Both variants import the same logic to avoid duplication,
- * since token generation is driver-agnostic — it only depends on the AWS SDK.
- *
- * @module token-generator
- */
-
-import { DsqlSigner } from "@aws-sdk/dsql-signer";
-
-/**
- * Options for generating an IAM authentication token.
- *
- * @property endpoint - The Aurora DSQL cluster hostname
- *   (e.g., "foo0bar1baz2quux3quuux4.dsql.us-east-1.on.aws")
- * @property region - AWS region for the cluster. Defaults to the `AWS_REGION`
- *   environment variable, then falls back to "us-east-1".
- * @property isAdmin - Whether to generate an admin-level token. Admin tokens
- *   use `getDbConnectAdminAuthToken()` and grant full privileges; non-admin
- *   tokens use `getDbConnectAuthToken()` for restricted roles.
- */
-export interface TokenOptions {
-  endpoint: string;
-  region?: string;
-  isAdmin?: boolean;
-}
-
-/**
- * Generates a short-lived IAM authentication token for Aurora DSQL.
- *
- * The token is produced by `@aws-sdk/dsql-signer` using the default AWS
- * credential chain (environment variables, shared credentials file, IAM role,
- * etc.). The caller uses this token as the PostgreSQL connection password.
- *
- * @param options - Token generation options (endpoint, region, admin flag)
- * @returns A short-lived authentication token string
- * @throws {Error} If token generation fails, with a message containing both
- *   the cluster endpoint and the underlying failure reason
- *
- * @example
- * ```ts
- * const token = await generateToken({
- *   endpoint: "my-cluster.dsql.us-east-1.on.aws",
- *   isAdmin: true,
- * });
- * ```
- */
-export async function generateToken(options: TokenOptions): Promise<string> {
-  const region =
-    options.region ?? Deno.env.get("AWS_REGION") ?? "us-east-1";
-
-  const signer = new DsqlSigner({
-    hostname: options.endpoint,
-    region,
-  });
-
-  try {
-    const token = options.isAdmin
-      ? await signer.getDbConnectAdminAuthToken()
-      : await signer.getDbConnectAuthToken();
-    return token;
-  } catch (error) {
-    throw new Error(
-      `Failed to generate IAM token for endpoint '${options.endpoint}': ${
-        error instanceof Error ? error.message : String(error)
-      }`,
-    );
-  }
-}
